HDAWG driver updates
#### **The following changes are made in `.ini` file:**
##### **1) New parameters and options are added to `Setup` section:**
* Revisions - Data Server
* Revisions - Firmware
* Revisions - FPGA
* Preset - Factory Reset
* PQSC - Connect to PQSC
##### **2) Following options are hidden when connected to PQSC:**
* Device - Reference Clock
* Digital I/O - Interface
* Digital I/O - Mode
* Digital I/O - Output
* Digital I/O - Drive
##### **3) `Trigger Mode` of `Seqeuncer` sections are updated:**
* `External Trigger` is replaced with `Receive Trigger`
* `ZSync Trigger` trigger mode is added
##### **4) New options are added to `DIO` section:**
* Digital I/O - Interface
* Digital I/O - Mode
##### **5) Following options are hidden when ZSYNC trigger chosen:**
* Sequencer - Strobe Index
* Sequencer - Strobe Slope
* Sequencer - Valid Index
* Sequencer - Valid Polarity
* Sequencer - Codeword Mask
* Sequencer - Codeword Shift
##### **6) "Rerun" option of sequencer is renamed to "Single"**
#### **The following changes are made in `.py` file:**
##### **1) Docstring and licence info added to beginning**
##### **2) `performOpen` method updated:**
The revision parameters are read from the device and the Labber
controller is updated in the beginning, right after connection is
established.
##### **3) `factory_reset` is called in `performSetValue`**
##### **4) `connect_to_pqsc` method is added:**
This method calls the necessary functions from zhinst-toolkit.
##### **5) AWGs are not started at the end of `performSetValue`**
Starting the AWGs automatically if they are in "Send Trigger" mode
causes the AWGs to be started two times when it is used to trigger
other devices. Therefore, these lines are removed.
##### **6) `performArm` method is updated:**
`External Trigger` is replaced with `Receive Trigger` and
`ZSync Trigger` is added to the arm condition.
##### **7) `get_node_value` is added to `performGetValue`**
`External Trigger` is replaced with `Receive Trigger` and
`ZSync Trigger` is added to the arm condition.
##### **8) `update_labber_controller` method is added:**
This method is used to read a quantity from the device and update the
Labber controller with the current value obtained.
##### **9) `update_sequencers` method is updated:**
The sequencer parameters should be set when the trigger mode
selected is `ZSync Trigger` even when the sequence type is `None`. This
is necessary when the user wants to use the HDAWG as a bridge between
the PQSC and UHFQA actually using the AWG sequencer of the HDAWG.