Refactoring (#18)
Adapt labber drivers to new toolkit 0.3

This is a major refactoring of the Labber drivers and breaks the API in many ways. It is however necessary since the old drivers did not scale with the fast pace our software and instruments evolve.

Before you upgrade read the documentation and see how the changes affect your workflow. (The old driver will continue to work but will not be actively maintained!)

The refactored driver has the following features:

* Introduce python package zhint-labber which is able to automatically generate
* the Labber driver for each Zurich Instruments device (including HF2).
* Easy to access CLI for zhinst-labber.
* Node naming aligned with LabOne.
* Automatic support for all nodes available on the device.
* Enhanced logging functionality.
* Waveform Processor that is able to convert native AWG Waveforms into its parts.
* Add DAQ, Sweeper and SHFQA sweeper modules

Co-authored-by: Markus Heikkinen <markus.heikkinen@zhinst.com>